+++ CHANGELOG.md
@@ -21,9 +21,11 @@ Status: Available for use
 ### Added
 
 - Allow entrypoint suppression to be set in the configuration file - MINOR
+- Add exec mode for generating images - MINOR
 
 ### Fixed
 - Fix up old dependencies; move to thiserror/anyhow instead of failure - MINOR
+- Publish tags to crates.io - PATCH
 
 ## [0.6.2] - 2021-03-03